@BrunoViniciusB NavBar e Search inseridos na tela de chat1
